requirements.txt
# TallyConnect - Modern Tally Sync Platform
# Python Dependencies

# Core Dependencies
pyodbc>=5.0.0
# ODBC connector for Tally database

# GUI Framework (Built-in with Python)
# tkinter - No separate install needed

# Standard Library (No install needed)
# sqlite3, threading, datetime, os, sys

# Build Dependencies (Development Only)
pyinstaller>=6.0.0
# For creating standalone executable

# System Tray Support
pystray>=0.19.0
Pillow>=10.0.0
# For system tray icons (Portal and Main app)

# Environment Variables (Phase 2)
python-dotenv>=1.0.0
# For loading configuration from .env file

# Scheduled Maintenance (Phase 3)
schedule>=1.2.0
# For scheduled maintenance tasks (vacuum, log cleaning, health checks)

# Redis Caching (Phase 4: Optional - only if REDIS_ENABLED=true)
redis>=5.0.0
# For high-performance caching (falls back to in-memory if not available)

# Encryption (Phase 5: Security & Validation)
cryptography>=41.0.0
# For AES-256 encryption of sensitive data (Fernet)

# Testing (Phase 6: Code Quality)
pytest>=7.4.0
pytest-cov>=4.1.0
# For unit testing and code coverage

